New changes commited

+++ demo/src/main/java/com/example/demo/controller/EmployeeController.java
@@ -27,7 +27,7 @@ public class EmployeeController {
     }
 
     @GetMapping("/{id}")
-    public Employee getEmployeeById(@PathVariable Long id) {
+    public Employee getEmployeeById(@PathVariable int id) {
         return employeeService.getEmployeeById(id);
     }
 
@@ -37,7 +37,7 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteEmployee(@PathVariable Long id) {
+    public String deleteEmployee(@PathVariable int id) {
         employeeService.deleteEmployee(id);
         return "Employee with ID " + id + " deleted.";
     }
